split common, eth and btc code
